backend/requirements.txt
Django==5.1.4 # Web framework
psycopg[binary]==3.3.2 # PostgreSQL database adapter
django-environ==0.12.0 # Environment variable management
djangorestframework==3.16.1 # REST framework for Django
django-cors-headers==4.9.0 # CORS headers management
Pillow==12.0.0 # Image processing library
djangorestframework-simplejwt==5.5.1 # JWT authentication for DRF
django-ratelimiter==0.2.2 # Rate limiting for Django public API
drf-yasg==1.21.11 # Automated generation of real Swagger/OpenAPI 2.0 schemas from DRF code
channels==4.3.2 # Asynchronous support for Django
channels-redis==4.2.1 # Redis backend for Django Channels
daphne==4.2.1 # ASGI server for Django Channels
redis==5.3.0 # Redis client for Python